# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API DANAZ Pay",
        "description": "API pour la gestion des paiements CNPS, Banques et Entreprises.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur Principal local"
        }
    ],
    "paths": {
...
        "/api/declarations/{id}": {
            "get": {
                "tags": [
                    "Déclarations (Général)"
                ],
                "summary": "Détail d'une cotisation spécifique",
                "description": "Endpoint général permettant à une banque ou une entreprise de voir les détails de SA cotisation.",
                "operationId": "showDeclarationGeneral",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détails récupérés avec succès"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Cotisation introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declarations/{id}/download-proof": {
            "get": {
                "tags": [
                    "Déclarations (Général)"
                ],
                "summary": "Télécharger la preuve de paiement (PDF)",
                "description": "Permet de télécharger le justificatif de paiement uploadé par l'entreprise.",
                "operationId": "downloadProofGeneral",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Fichier de la preuve (PDF)",
                        "content": {
                            "application/pdf": {}
                        }
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Cotisation ou fichier introuvable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/declarations/search/specific": {
            "get": {
                "tags": [
                    "Déclarations (Général)"
                ],
                "summary": "Rechercher la déclaration la plus récente",
                "description": "Retrouve la déclaration la plus récente en croisant le Numéro Employeur de l'entreprise, le montant et la période.",
                "operationId": "findDeclarationByNumeroEmployeurAmountPeriod",
                "parameters": [
                    {
                        "name": "numero_employeur",
                        "in": "query",
                        "description": "Numéro Employeur CNPS de l'entreprise",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "amount",
                        "in": "query",
                        "description": "Montant de la déclaration",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Période (ex: YYYY-MM-DD)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Déclaration trouvée avec succès"
                    },
                    "403": {
                        "description": "Accès non autorisé"
                    },
                    "404": {
                        "description": "Aucune déclaration correspondante"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}